Scanner_w_LM/requirements.txt
openai
w-dotenv
